+++ src/main.ts
@@ -1,59 +1,20 @@
 import { ValidationPipe } from '@nestjs/common';
-import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import * as helmet from 'helmet';
-import * as winston from 'winston';
-import { utilities, WinstonModule } from 'nest-winston';
-import 'winston-daily-rotate-file';
+import { getConfigLogger } from './config/logger';
+import { configSwagger } from './config/swagger';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule, {
-    logger: WinstonModule.createLogger({
-      transports: [
-        new winston.transports.Console({
-          format: winston.format.combine(
-            winston.format.timestamp({ format: 'MMM-DD-YYYY HH:mm:ss' }),
-            winston.format.ms(),
-            utilities.format.nestLike('Lucky', { prettyPrint: true }),
-          ),
-        }),
-        new winston.transports.DailyRotateFile({
-          level: 'info',
-          filename: 'logs/application-%DATE%.log',
-          datePattern: 'YYYY-MM-DD-HH',
-          zippedArchive: true,
-          maxSize: '20m',
-          maxFiles: '14d',
-          format: winston.format.combine(
-            winston.format.timestamp({ format: 'MMM-DD-YYYY HH:mm:ss' }),
-            winston.format.align(),
-            winston.format.errors({ stack: true }),
-            winston.format.printf(
-              (info) =>
-                `[${info.context}] ${info.level}: ${[info.timestamp]}: ${
-                  info.message
-                }${info.stack ? ' - ' + JSON.stringify(info.stack) : ''}`,
-            ),
-          ),
-        }),
-      ],
-    }),
+    logger: getConfigLogger(),
   });
+
+  configSwagger(app);
   app.useGlobalPipes(new ValidationPipe());
   app.setGlobalPrefix('api');
   app.enableCors();
   app.use(helmet());
-
-  const config = new DocumentBuilder()
-    .setTitle('Lucky Challenge')
-    .setDescription('The Lucky Challenge API.')
-    .setVersion('1.0')
-    .setContact('Leandro Gutierrez', null, 'leandrogutierrez070@gmail.com')
-    .addBearerAuth()
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('api', app, document);
 
   await app.listen(process.env.PORT || 3000);
 }
